# Simplify subtree merging.